Slightly rewrote Sprees and Combos.
This is written in a more elegant way. Also, HUDMessages in a very primitive way are added, only clientside. I'll have to rewrite the protocol for including HUDMessages from the server.

+++ common/p_ctf.h
@@ -55,7 +55,7 @@ struct flagdata
 
 	// Integer representation of WHO has each flag (player id)
 	byte flagger;
-	int	pickup_time;
+	long pickup_time;
 
 	// Flag locations
 	int x, y, z;
